+++ src/containers/BooksForm.js
@@ -1,12 +1,16 @@
 import { useState } from 'react';
+import { connect } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
+import PropTypes from 'prop-types';
+import { createBookAction } from '../actions/index';
 
-const BooksForm = () => {
+const BooksForm = ({ createBook }) => {
   const Category = ['Action', 'Biography', 'History', 'Horror', 'Kids', 'Learning', 'Sci-Fi'];
   const [state, setState] = useState({ title: '', category: '' });
   const handleChange = ({ target: { name, value } }) => {
     setState({ ...state, [name]: value });
   };
+
   const { title, category } = state;
   const handleSubmit = (e) => {
     e.preventDefault();
@@ -23,13 +27,14 @@ const BooksForm = () => {
       <h2>Add Book here</h2>
       <form onSubmit={handleSubmit}>
         <label htmlFor="title">Book title:</label>
-        <input type="text" onChange={handleChange} id="title" name="title" placeholder="Title of the book" required />
+        <input type="text" value={state.title || ''} onChange={handleChange} id="title" name="title" placeholder="Title of the book" required />
         <select
           onChange={handleChange}
           name="category"
+          value={state.category}
           required
         >
-          <option disabled value>
+          <option value="default">
             Select a category
           </option>
           {Category.map(category => (
@@ -44,4 +49,10 @@ const BooksForm = () => {
   );
 };
 
-export default BooksForm;
+BooksForm.propTypes = {
+  createBook: PropTypes.func.isRequired,
+};
+
+const mapDispatchToProps = dispatch => ({ createBook: book => dispatch(createBookAction(book)) });
+
+export default connect(null, mapDispatchToProps)(BooksForm);
